+++ History.md
@@ -5,10 +5,14 @@ Terraform module to buld a simple RKE2 cluster.
 ### v0.1.0
 The first release.
 
-### v0.2.0
+### v0.2
 Add
 - Server Taints Option
 - Server Disable Option
 - Token Auto Creation
 - Allocate Public IP fix when server.single is true
 - Documents
+
+## v0.3
+- Auto delete terminated agent node from cluster.
+- Restore automation.